Film index works, database can be migrated

+++ Filminurk/Filminurk/Models/Movies/MoviesIndexViewModel.cs
@@ -7,7 +7,7 @@ namespace Filminurk.Models.Movies
         public Guid ID { get; set; }
         public string Title { get; set; }
         public DateOnly FirstPublished { get; set; }
-        public decimal? CurrentRating { get; set; }
+        public double? CurrentRating { get; set; }
 
         /* 2 õpilase valitud andmetüübi */
